refactor and new structure
